Get kafka and zeromq extra tests passing
