+++ setup.py
@@ -64,11 +64,12 @@ setup(
             'setuptools>=50.3.1',
         ],
         'zeromq': [
-            'pyzmq',
+            'pyzmq>=19.0.2',
             'msgpack-python>=0.4'
         ],
         'kafka': [
-            'kafka-python>=1.4.0'
+            'kafka-python>=1.4.3',
+            'twisted>=20.3.0',
         ],
         'distributed': [
             'Twisted'
