+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = pre-commit,min,min-hbase,min-logging,min-redis,min-scrapy,min-sql,min-all,py39,py310,py311,py312,py313,hbase,logging,redis,scrapy,sql,all
+envlist = pre-commit,min,min-hbase,min-kafka,min-logging,min-redis,min-scrapy,min-sql,min-zeromq,min-all,py39,py310,py311,py312,py313,hbase,kafka,logging,redis,scrapy,sql,zeromq,all
 # TODO: add min-* and * for each extra and for all at once.
 minversion = 1.8
 
@@ -10,6 +10,7 @@ setenv =
     BOTO_CONFIG = /tmp/nowhere
     FRONTERA_TESTING = true
 deps =
+    flaky
     pytest
     pytest-cov
 commands =
@@ -17,6 +18,11 @@ commands =
 
 [testenv:hbase]
 extras = hbase
+
+[testenv:kafka]
+extras = kafka
+# https://github.com/dpkp/kafka-python/issues/2412
+basepython = python3.11
 
 [testenv:logging]
 extras = logging
@@ -30,13 +36,18 @@ extras = scrapy
 [testenv:sql]
 extras = sql
 
+[testenv:zeromq]
+extras = zeromq
+
 [testenv:all]
 extras =
     hbase
+    kafka
     logging
     redis
     scrapy
     sql
+    zeromq
 
 [testenv:min]
 basepython = python3.9
@@ -54,6 +65,13 @@ deps =
     happybase==1.2.0
     msgpack-python==0.4
     setuptools==50.3.1
+
+[testenv:min-kafka]
+basepython = {[testenv:min]basepython}
+deps =
+    {[testenv:min]deps}
+    kafka-python==1.4.3
+    twisted==20.3.0
 
 [testenv:min-logging]
 basepython = {[testenv:min]basepython}
@@ -83,15 +101,24 @@ deps =
     cachetools==0.4.0
     SQLAlchemy==1.0.0
 
+[testenv:min-zeromq]
+basepython = {[testenv:min]basepython}
+deps =
+    {[testenv:min]deps}
+    pyzmq==19.0.2
+    msgpack-python==0.4
+
 [testenv:min-all]
 basepython = {[testenv:min]basepython}
 deps =
     {[testenv:min]deps}
     {[testenv:min-hbase]deps}
+    {[testenv:min-kafka]deps}
     {[testenv:min-logging]deps}
     {[testenv:min-redis]deps}
     {[testenv:min-scrapy]deps}
     {[testenv:min-sql]deps}
+    {[testenv:min-zeromq]deps}
 
 [testenv:pre-commit]
 deps = pre-commit
